Add SpotifyStreamingHistoryEntry interface
spotify-processor.ts contains the logic for parsing through Spotify's
"Extended Streaming History" JSON files. The SpotifyStreamingHistoryEntry
interface matches the fields that are relevant for song entries (i.e.,
no podcast or audiobook information is tracked). In the future, we will
want functionality to process streaming data in an online fashion that
is not from the "Extended Streaming History" JSON dump, so a different
interface may be implemented for such streaming entries. This "Extended
Streaming History" upload should be though of as a one-time (optional)
action for the user.